+++ src/app/modals/research-modal/research-modal.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit, Inject } from '@angular/core';
 import { DatabaseService } from 'src/app/services/database.service';
-import { MAT_DIALOG_DATA } from '@angular/material/dialog';
-import { inject } from '@angular/core/testing';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { AngularFireStorage } from '@angular/fire/compat/storage';
+import { ResearchItem, FIREBASE_PATHS } from 'src/app/models/content.models';
 
 @Component({
   selector: 'app-research-modal',
@@ -15,7 +15,7 @@ export class ResearchModalComponent implements OnInit {
   img_name: string = "";
   title: string = "";
   length: number;
-  researchData: any;
+  researchData: ResearchItem[];
   index: number;
   type: string;
   image: string = "";
@@ -23,6 +23,7 @@ export class ResearchModalComponent implements OnInit {
   constructor(
     private databaseService: DatabaseService,
     private storage: AngularFireStorage,
+    private dialogRef: MatDialogRef<ResearchModalComponent>,
     @Inject(MAT_DIALOG_DATA) public data: any
   ) { }
 
@@ -37,7 +38,7 @@ export class ResearchModalComponent implements OnInit {
       this.index = this.data.index;
       this.content = this.data.data[this.index].content
       this.title = this.data.data[this.index].title
-      this.length = this.data.length; 
+      this.length = this.data.length;
       this.type = this.data.type;
       this.image = this.data.data[this.index].img_name;
     }
@@ -45,18 +46,19 @@ export class ResearchModalComponent implements OnInit {
 
   async addResearch() {
     if (this.type == 'add') {
-      this.researchData.splice(0,0, {content: this.content, img_name: this.image, title: this.title})
-      await this.databaseService.addResearchData(`research/research/`, this.researchData);
+      this.researchData.splice(0, 0, { content: this.content, img_name: this.image, title: this.title })
+      await this.databaseService.addResearchData(FIREBASE_PATHS.RESEARCH, this.researchData);
       alert("Successfully added Data");
     }
 
     if (this.type == 'edit') {
-      this.researchData[this.index] = {content: this.content, img_name: this.image, title: this.title}
+      this.researchData[this.index] = { content: this.content, img_name: this.image, title: this.title }
       console.log(this.researchData)
-      await this.databaseService.addResearchData(`research/research/`, this.researchData);
+      await this.databaseService.addResearchData(FIREBASE_PATHS.RESEARCH, this.researchData);
       alert("Successfully edited Data");
     }
-    
+
+    this.dialogRef.close();
   }
 
   async onFileChange(event: any) {
@@ -70,15 +72,18 @@ export class ResearchModalComponent implements OnInit {
   }
 
   async removeImage() {
+    if (!this.image) {
+      return;
+    }
     if (confirm('Are you sure you want to remove this image?')) {
       console.log('Removing image')
       this.storage.refFromURL(this.image).delete();
       this.image = "";
-      this.researchData[this.index] = {content: this.content, img_name: this.image, title: this.title}
-      await this.databaseService.addResearchData(`research/research/`, this.researchData);
+      if (this.type === 'edit' && this.index != null) {
+        this.researchData[this.index] = { content: this.content, img_name: this.image, title: this.title }
+        await this.databaseService.addResearchData(FIREBASE_PATHS.RESEARCH, this.researchData);
+      }
     }
-    
-
   }
 
 }
